test: move token counting tests
